Implemented delete user in DB; Junit tests for UserService.

+++ src/main/java/com/herokuapp/prop3rty/service/UserService.java
@@ -20,8 +20,13 @@ public class UserService {
 		return dao.update(user);	
 	}
 	
-	public boolean delete(User user) {
-		return dao.delete(user);
+	public boolean delete(String email) {
+		User user = dao.searchByEmail(email);
+		if (user == null) {
+			return false;
+		} else {
+			return dao.delete(user);
+		}
 	}
 	
 }
